requirements.txt
crewai>=0.86.0,<1.0.0
streamlit>=1.39.0
chromadb>=0.5.0
openai>=1.50.0
python-dotenv>=1.0.1
pypdf>=5.0.0
pandas>=2.0.0
